+++ packages/video_player_videohole/lib/src/video_player_tizen.dart
@@ -35,20 +35,16 @@ class VideoPlayerTizen extends VideoPlayerPlatform {
       case DataSourceType.asset:
         message.asset = dataSource.asset;
         message.packageName = dataSource.package;
-        break;
       case DataSourceType.network:
         message.uri = dataSource.uri;
         message.formatHint = _videoFormatStringMap[dataSource.formatHint];
         message.httpHeaders = dataSource.httpHeaders;
         message.drmConfigs = dataSource.drmConfigs?.toMap();
         message.playerOptions = dataSource.playerOptions;
-        break;
       case DataSourceType.file:
         message.uri = dataSource.uri;
-        break;
       case DataSourceType.contentUri:
         message.uri = dataSource.uri;
-        break;
     }
 
     final PlayerMessage response = await _api.create(message);