+++ packages/wearable_rotary/lib/wearable_rotary.dart
@@ -1,4 +1,2 @@
-library wearable_rotary;
-
 export 'src/rotary_scroll_controller.dart';
 export 'src/wearable_rotary_base.dart';