+++ packages/battery_plus/CHANGELOG.md
@@ -1,3 +1,7 @@
+## NEXT
+
+* Update minimum Flutter and Dart version to 3.13 and 3.1.
+
 ## 1.1.2
 
 * Update battery_plus to 4.0.1.