+++ packages/video_player_videohole/CHANGELOG.md
@@ -1,3 +1,8 @@
+## NEXT
+
+* Fix new lint warnings.
+* Update minimum Flutter and Dart version to 3.13 and 3.1.
+
 ## 0.5.1
 
 * Free response buffer after install key.