+++ packages/wearable_rotary/CHANGELOG.md
@@ -1,3 +1,8 @@
+## NEXT
+
+* Fix new lint warnings.
+* Update minimum Flutter and Dart version to 3.13 and 3.1.
+
 ## 2.0.2
 
 * Adds namespace to `build.gradle` for compatibility with Gradle 8